ranked search on search service enabled. also fixed flaw - after cita…

+++ SearchService/src/main/java/com/ynm/searchservice/Model/ResearchPaper.java
@@ -20,5 +20,5 @@ public class ResearchPaper {
     private String visibility;
     private Integer ownerId;
     private java.sql.Date createdAt;
-    private Integer metric;
+    private Float metric;
 }
